+++ lib/src/parser/repeater/limited.dart
@@ -5,7 +5,7 @@ import 'repeating.dart';
 /// of its delegate and that requires the input to be completed with a specified
 /// parser 'limit'. Subclasses provide repeating behavior as typically seen in
 /// regular expression implementations (non-blind).
-abstract class LimitedRepeatingParser<R> extends RepeatingParser<R> {
+abstract class LimitedRepeatingParser<R> extends RepeatingParser<R, List<R>> {
   LimitedRepeatingParser(Parser<R> delegate, this.limit, int min, int max)
       : super(delegate, min, max);