+++ lib/src/parser/repeater/possessive.dart
@@ -47,7 +47,7 @@ extension PossessiveRepeatingParserExtension<T> on Parser<T> {
 
 /// A greedy parser that repeatedly parses between 'min' and 'max' instances of
 /// its delegate.
-class PossessiveRepeatingParser<R> extends RepeatingParser<R> {
+class PossessiveRepeatingParser<R> extends RepeatingParser<R, List<R>> {
   PossessiveRepeatingParser(super.parser, super.min, super.max);
 
   @override