+++ test/regression_test.dart
@@ -1,3 +1,4 @@
+// ignore_for_file: deprecated_member_use_from_same_package
 import 'dart:math';
 
 import 'package:petitparser/petitparser.dart';